added popups to every delete operation
