+++ frontend/src/lib/components/ThoughtItem.svelte
@@ -1,4 +1,7 @@
 <script>
+    import { openPopup } from "../popup";
+    import DeleteConfirmationPopup from "./DeleteConfirmationPopup.svelte";
+
     export let title;
     export let date;
     export let iconType;
@@ -13,6 +16,7 @@
       icon = "neutral";
     }
 </script>
+<DeleteConfirmationPopup name="thought" />
 <div class="bg-white rounded-xl py-2 px-3 flex flex-row w-full justify-between items-center" style="box-shadow: rgba(0, 0, 0, 0.15) 2.4px 2.4px 3.2px">
   <a href={`thoughts/${thoughtId}`} class="flex flex-row gap-4">
     <div>
@@ -24,8 +28,8 @@
     </div>
   </a>
   <div>
-    <form class="w-full rounded-lg p-1" method="POST" action={`/moods/thoughts/${thoughtId}?/deleteThought`}>
-      <button type="submit"><img src="../../src/img/trashIcon.png" alt="trash" class="h-8"></button>
+    <form id="deleteForm" class="w-full rounded-lg p-1" method="POST" action={`/moods/thoughts/${thoughtId}?/deleteThought`}>
+      <button on:click|preventDefault={openPopup}><img src="../../src/img/trashIcon.png" alt="trash" class="h-8"></button>
     </form>
   </div>
 </div>
