+++ frontend/src/routes/categories/[categoryId]/edit/+page.svelte
@@ -67,11 +67,11 @@ $showElement = false;
                         >
                             <a
                                 href="/categories"
-                                class="px-5 py-2 rounded-lg mt-3 font-bold"
+                                class="px-5 py-2 rounded-lg mt-3 font-bold text-white"
                                 style="background-color: #B4B4B4">Cancel</a
                             >
                             <button
-                                class="px-5 py-2 rounded-lg mt-3 font-bold"
+                                class="px-5 py-2 rounded-lg mt-3 font-bold text-white"
                                 type="submit"
                                 style="background-color: #9B9DD1">Submit</button
                             >
