+++ frontend/src/routes/categories/[categoryId]/habits/[habitId]/edit/+page.svelte
@@ -89,11 +89,11 @@ imgExtraPath="../../../"
                         >
                             <a
                                 href="/categories/{data.category}/habits"
-                                class="px-5 py-2 rounded-lg mt-3 font-bold"
+                                class="px-5 py-2 rounded-lg mt-3 font-bold text-white"
                                 style="background-color: #B4B4B4">Cancel</a
                             >
                             <button
-                                class="px-5 py-2 rounded-lg mt-3 font-bold"
+                                class="px-5 py-2 rounded-lg mt-3 font-bold text-white"
                                 type="submit"
                                 style="background-color: #9B9DD1">Submit</button
                             >
